[신윤식] - 어렵다 파라미터 html 컨트롤러 DB 이름 수정 및 데이터 속성 사용으로 postId 마다 각개적으로 활용

+++ src/main/java/com/andy/playground/likes/LikesRestController.java
@@ -24,8 +24,8 @@ public class LikesRestController {
 
     @PostMapping("/like")
     public Map<String, String> createLike(
-            @RequestParam("postId") int postId
-            , HttpSession session) {
+            @RequestParam("postId") long postId, 
+            HttpSession session) {
 
         Map<String, String> resultMap = new HashMap<>();
